booking api

+++ migrations/20210818185821-create_table_address.js
@@ -34,11 +34,11 @@ module.exports = {
         allowNull: true,
       },
       longtitude: {
-        type: Sequelize.BIGINT(30),
+        type: Sequelize.FLOAT(10, 6),
         allowNull: true,
       },
       latitude: {
-        type: Sequelize.BIGINT(30),
+        type: Sequelize.FLOAT(10, 6),
         allowNull: true,
       },
